creating new folder, css file for component NavBar
