+++ app/components/layout.module.css
@@ -14,35 +14,6 @@
   margin: 3rem 0 0;
 }
 
-.navBar {
-  background-color: #b7997b;
-  text-decoration: none;
-  color: black;
-  width: 100%;
-  height: 80px;
-  overflow: hidden;
-}
-
-.navBarLink {
-  text-decoration: none;
-  font-size: 18px;
-  padding: 50px 50px 50px 50px;
-  color: black;
-  cursor: pointer;
-}
-.navBarLink:after,
-.navBarLink:before {
-  transition: all 0.2.s;
-}
-.navBarLink:hover {
-  transition-duration: 0.1s;
-  letter-spacing: 3px;
-}
-
-.menuLeft {
-  float: left;
-}
-
-.menuRight {
-  float: right;
+.nameStyles {
+  font-style: italic;
 }
